separação da responsabilidade de Slide e SlideItem de desenharem-se
