+++ src/main/java/br/upe/enenhariasoftware/psw/jabberpoint/controller/KeyController.java
@@ -10,7 +10,6 @@
 package br.upe.enenhariasoftware.psw.jabberpoint.controller;
 
 import br.upe.enenhariasoftware.psw.jabberpoint.model.Presentation;
-import br.upe.enenhariasoftware.psw.jabberpoint.view.SlideViewerFrame;
 
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
