+++ src/main/java/br/upe/enenhariasoftware/psw/jabberpoint/model/SlideItem.java
@@ -29,6 +29,4 @@ public abstract class SlideItem {
 		return level;
 	}
 
-	public abstract Rectangle getBoundingBox(Graphics g, ImageObserver observer, float scale, Style style);
-
 }
