+++ src/main/java/br/upe/enenhariasoftware/psw/jabberpoint/model/TextItem.java
@@ -35,55 +35,6 @@ public class TextItem extends SlideItem {
 		return text == null ? "" : text;
 	}
 
-	public AttributedString getAttributedString(Style style, float scale) {
-		AttributedString attrStr = new AttributedString(getText());
-
-		attrStr.addAttribute(TextAttribute.FONT, style.getFont(scale), 0, text.length());
-
-		return attrStr;
-	}
-
-	public Rectangle getBoundingBox(Graphics g, ImageObserver observer, float scale, Style myStyle) {
-		List<TextLayout> layouts = getLayouts(g, myStyle, scale);
-
-		int xsize = 0;
-		int ysize = (int) (myStyle.leading * scale);
-
-		for (TextLayout layout : layouts) {
-			Rectangle2D bounds = layout.getBounds();
-
-			if (bounds.getWidth() > xsize) {
-				xsize = (int) bounds.getWidth();
-			}
-
-			if (bounds.getHeight() > 0) {
-				ysize += bounds.getHeight();
-			}
-			ysize += layout.getLeading() + layout.getDescent();
-		}
-
-		return new Rectangle((int) (myStyle.indent * scale), 0, xsize, ysize);
-	}
-
-	private List<TextLayout> getLayouts(Graphics g, Style s, float scale) {
-		List<TextLayout> layouts = new ArrayList<>();
-
-		AttributedString attrStr = getAttributedString(s, scale);
-		Graphics2D g2d = (Graphics2D) g;
-
-		FontRenderContext frc = g2d.getFontRenderContext();
-		LineBreakMeasurer measurer = new LineBreakMeasurer(attrStr.getIterator(), frc);
-
-		float wrappingWidth = (Slide.WIDTH - s.indent) * scale;
-
-		while (measurer.getPosition() < getText().length()) {
-			TextLayout layout = measurer.nextLayout(wrappingWidth);
-			layouts.add(layout);
-		}
-
-		return layouts;
-	}
-
 	@Override
 	public String toString() {
 		return "TextItem[" + getLevel() + "," + getText() + "]";
